Also prepare the filesystem when a command is given
start.sh is run when no command is given when starting a container.

But when the makefiles run the container, with a composer install
command, we also need to make sure the filesystem got the correct ACLs.

To do so, we now prepare the filesystem in a new prepare.sh entrypoint.

We use an additional prepare.sh, and not entry.sh, because it allow the
projects to extends this, without rewriting entry.sh.

This also clean and improve the shell code.